+++ simpleraytracer/camera.h
@@ -110,7 +110,7 @@ public:
 	virtual SampleLensResults sampleLens(
 		const Vec2f& imagepos,
 		const SamplePair& samples,
-		double time,
+		float time,
 		bool vignetting,
 		Real recip_normed_image_rect_area,
 		bool compute_alpha_weight
@@ -122,23 +122,23 @@ public:
 
 	virtual PDType sensorPDF() const = 0; // used
 
-	virtual void sampleLensPos(const SamplePair& samples, double time, Vec3Type& pos_os_out, Vec3Type& pos_ws_out) const = 0; // used
+	virtual void sampleLensPos(const SamplePair& samples, float time, Vec3Type& pos_os_out, Vec3Type& pos_ws_out) const = 0; // used
 
-	virtual const Vec3Type lensExitDir(const Vec3Type& sensorpos_os, const Vec3Type& lenspos_os, double time) const = 0; // only used in old code.
+	virtual const Vec3Type lensExitDir(const Vec3Type& sensorpos_os, const Vec3Type& lenspos_os, float time) const = 0; // only used in old code.
 
-	virtual const Vec2f imCoordsForSensorPos(const Vec3Type& sensorpos_os, double time) const = 0;
+	virtual const Vec2f imCoordsForSensorPos(const Vec3Type& sensorpos_os, float time) const = 0;
 	virtual void sensorPosForImCoords(const Vec2f& imcoords, Vec3Type& pos_os_out) const = 0;
 
 
-	virtual const Vec4f getForwardsDirF(double time) const = 0;
+	virtual const Vec4f getForwardsDirF(float time) const = 0;
 
-	virtual const Vec4f getPosWS(double time) const = 0;
+	virtual const Vec4f getPosWS(float time) const = 0;
 	
 
 	virtual Real getExposureDuration() const = 0;
 
 
-	virtual const Vec3Type diffractRay(const SamplePair& samples, const Vec3Type& dir, const WavelengthSamples& wavelengths, float direction_sign, double time, SpectralVector& weights_out) const = 0;
+	virtual const Vec3Type diffractRay(const SamplePair& samples, const Vec3Type& dir, const WavelengthSamples& wavelengths, float direction_sign, float time, SpectralVector& weights_out) const = 0;
 
 
 	virtual void prepareForDiffractionFilter(int main_buffer_width, int main_buffer_height, int ssf_) = 0;
@@ -154,7 +154,7 @@ public:
 	virtual double imageHeightAtDistanceOne() const = 0;
 
 	// Used in IndigoDriver::traceRay() for picking.
-	virtual void getRayForImagePos(const Vec2f& image_coordinates, double time, Vec4f& pos_ws_out, Vec4f& dir_ws_out) const = 0;
+	virtual void getRayForImagePos(const Vec2f& image_coordinates, float time, Vec4f& pos_ws_out, Vec4f& dir_ws_out) const = 0;
 
 
 	virtual Real meanCurvature(const HitInfo& hitinfo) const { return 0; }
